rename button components to comply with naming scheme
