+++ web-ui/src/pages/index.tsx
@@ -6,9 +6,9 @@ import { Col, Row } from 'antd';
 import styles from './index.less';
 import { useEffect, useState } from 'react';
 import SearchResult from '@/components/search/search-results/search-result';
-import HugeButton from '@/components/input/buttons/huge-button/huge-button';
+import ButtonHuge from '@/components/input/button/button-huge/button-huge';
 import { FileSearchOutlined, VideoCameraAddOutlined } from '@ant-design/icons';
-import HugePreviewButton from '@/components/input/buttons/huge-preview-button/huge-preview-button';
+import ButtonHugeWithPreview from '@/components/input/button/button-huge-with-preview/button-huge-with-preview';
 import CourseList from '@/components/course/course-list/course-list';
 import ImageUpload from '@/components/image/image-upload/image-upload';
 import { Image } from '@/types/search';
@@ -57,7 +57,7 @@ export default function IndexPage() {
                 />
               </Col>
               <Col sm={24} md={8}>
-                <HugePreviewButton
+                <ButtonHugeWithPreview
                   icon={<FileSearchOutlined />}
                   title="Find a lecture"
                   subtitle="Find a lecture from the lectures kthGPT has already watched"
@@ -68,7 +68,7 @@ export default function IndexPage() {
                 />
               </Col>
               <Col sm={24} md={8}>
-                <HugeButton
+                <ButtonHuge
                   icon={<VideoCameraAddOutlined />}
                   title="Add a new lecture"
                   subtitle="It usually takes between 5 and 15 minutes"
